simple_worker/requirements.txt
bbox==0.9.2 
celery==5.2.6
matplotlib==3.5.1
opencv-python==4.5.4.60
opencv-python-headless==4.5.4.60
pandas==1.3.5
paramiko==2.11.0
pickle-mixin==1.0.2
python-bidi==0.4.2
PyYAML==6.0
scikit-image==0.19.2
scipy==1.7.3
paddleocr==2.5
paddlepaddle==2.3.0
python-dotenv==0.20.0
seaborn>=0.11.2
protobuf==3.20.1